+++ cmem_plugin_reason/plugin_reason.py
@@ -54,12 +54,12 @@ If there are classes `Person`, `Student` and `Professor`, such that `Person Disj
 Student, Professor` holds, the reasoner will infer `DisjointClasses: Student, Professor`.
 """
 
-# DATA_PROP_CHAR_DESC = """The reasoner will infer characteristics of data properties, i.e.
-# `Characteristics:` statements. For data properties, this only pertains to functionality.\n
-# If there are data properties `identifier` and `enrollmentNumber`, such that `enrollmentNumber
-# SubPropertyOf: identifier` and `identifier Characteristics: Functional` holds, the reasoner will
-# infer `enrollmentNumber Characteristics: Functional`.
-# """
+DATA_PROP_CHAR_DESC = """The reasoner will infer characteristics of data properties, i.e.
+`Characteristics:` statements. For data properties, this only pertains to functionality.\n
+If there are data properties `identifier` and `enrollmentNumber`, such that `enrollmentNumber
+SubPropertyOf: identifier` and `identifier Characteristics: Functional` holds, the reasoner will
+infer `enrollmentNumber Characteristics: Functional`.
+"""
 
 DATA_PROP_EQUIV_DESC = """The reasoner will infer axioms about the equivalence of data properties,
  i.e. `EquivalentProperties` statements.\n
@@ -91,13 +91,13 @@ assertions `John Facts: enrolledIn KnowledgeRepresentation` and `LeipzigUniversi
 KnowledgeRepresentation`,  the reasoner will infer `John Facts: enrolledIn LeipzigUniversity`.
 """
 
-# OBJECT_PROP_CHAR_DESC = """The reasoner will infer characteristics of object properties, i.e.
-# `Characteristics:` statements.\n
-# If there are object properties `enrolledIn` and `studentOf`, such that `enrolledIn
-# SubPropertyOf: studentOf` and `enrolledIn Characteristics: Functional` holds, the reasoner will
-# infer `studentOf Characteristics: Functional`. **Note: this inference does neither work in JFact
-# nor in HermiT!**
-# """
+OBJECT_PROP_CHAR_DESC = """The reasoner will infer characteristics of object properties, i.e.
+`Characteristics:` statements.\n
+If there are object properties `enrolledIn` and `studentOf`, such that `enrolledIn
+SubPropertyOf: studentOf` and `enrolledIn Characteristics: Functional` holds, the reasoner will
+infer `studentOf Characteristics: Functional`. **Note: this inference does neither work in JFact
+nor in HermiT!**
+"""
 
 OBJECT_PROP_EQUIV_DESC = """The reasoner will infer assertions about the equivalence of object
 properties, i.e. `EquivalentTo:` statements.\n
@@ -190,13 +190,13 @@ Person`.
             description=DISJOINT_DESC,
             default_value=False,
         ),
-        # PluginParameter(
-        #     param_type=BoolParameterType(),
-        #     name="data_property_characteristic",
-        #     label="Data property characteristics (Axiomatic triples)",
-        #     description=DATA_PROP_CHAR_DESC,
-        #     default_value=False,
-        # ),
+        PluginParameter(
+            param_type=BoolParameterType(),
+            name="data_property_characteristic",
+            label="Data property characteristics (Axiomatic triples)",
+            description=DATA_PROP_CHAR_DESC,
+            default_value=False,
+        ),
         PluginParameter(
             param_type=BoolParameterType(),
             name="equivalent_data_properties",
@@ -239,13 +239,13 @@ Person`.
             description=OBJECT_PROP_INV_DESC,
             default_value=False,
         ),
-        # PluginParameter(
-        #     param_type=BoolParameterType(),
-        #     name="object_property_characteristic",
-        #     label="Object property characteristic (Axiomatic triples)",
-        #     description=OBJECT_PROP_CHAR_DESC,
-        #     default_value=False,
-        # ),
+        PluginParameter(
+            param_type=BoolParameterType(),
+            name="object_property_characteristic",
+            label="Object property characteristic (Axiomatic triples)",
+            description=OBJECT_PROP_CHAR_DESC,
+            default_value=False,
+        ),
         PluginParameter(
             param_type=BoolParameterType(),
             name="sub_object_property",
@@ -321,16 +321,13 @@ class ReasonPlugin(WorkflowPlugin):
         disjoint_classes: bool = False,
         sub_object_property: bool = False,
         equivalent_object_property: bool = False,
-        #   This axiom generator does not yield any results.
-        #   Issue: https://github.com/eccenca/cmem-plugin-reason/issues/10
-        #   object_property_characteristic: bool = False,
+        object_property_characteristic: bool = False,
         object_property_domain: bool = False,
         object_property_range: bool = False,
         inverse_object_properties: bool = False,
         sub_data_property: bool = False,
         equivalent_data_properties: bool = False,
-        #   Removed because the object property counterpart does not work
-        #   data_property_characteristic: bool = False,
+        data_property_characteristic: bool = False,
         validate_profile: bool = False,
         import_ontology: bool = True,
         import_result: bool = False,
@@ -342,14 +339,14 @@ class ReasonPlugin(WorkflowPlugin):
             "SubClass": sub_class,
             "EquivalentClass": equivalent_class,
             "DisjointClasses": disjoint_classes,
-            # "DataPropertyCharacteristic": data_property_characteristic,
+            "DataPropertyCharacteristic": data_property_characteristic,
             "EquivalentDataProperties": equivalent_data_properties,
             "SubDataProperty": sub_data_property,
             "ClassAssertion": class_assertion,
             "PropertyAssertion": property_assertion,
             "EquivalentObjectProperty": equivalent_object_property,
             "InverseObjectProperties": inverse_object_properties,
-            # "ObjectPropertyCharacteristic": object_property_characteristic,
+            "ObjectPropertyCharacteristic": object_property_characteristic,
             "SubObjectProperty": sub_object_property,
             "ObjectPropertyRange": object_property_range,
             "ObjectPropertyDomain": object_property_domain,